refactor(server): split worker-api into per-concern route modules
Break the 3270-line worker-api.ts into a barrel plus focused modules
(poll, run-lifecycle, device-management, device-auth-profiles, device-feeds,
auth-runs, watcher-trigger, shared). The exported entrypoint and every route,
auth check, and payload are unchanged; security-sensitive paths moved verbatim.
Validated against the worker-api integration tests.
